+++ django_webix_sender/templates/django_webix_sender/invoices.js
@@ -1,9 +1,9 @@
 {% load static i18n verbose_name field_type %}
 
 {% block content %}
-webix.ui([], $$('content_right'));
+webix.ui([], $$('{{ webix_container_id }}'));
 
-$$("content_right").addView({
+$$("{{ webix_container_id }}").addView({
     rows: [
         {
             view: "toolbar",
@@ -12,16 +12,16 @@ $$("content_right").addView({
                     id: 'send_method',
                     view: 'combo',
                     value: "{{ request.GET.send_method }}",
-                    label: "{% trans 'Send method' %}",
+                    label: "{{_("Send method")|escapejs}}",
                     labelWidth: 130,
                     labelAlign: 'right',
                     options: [
                         {id: "", value: "", $empty: true},
                         {% for send_method in send_methods %}
-                        {
-                            id: "{{ send_method.key|safe|escapejs }}",
-                            value: "{{ send_method.value|safe|escapejs }}"
-                        },
+                            {
+                                id: "{{ send_method.key|safe|escapejs }}",
+                                value: "{{ send_method.value|safe|escapejs }}"
+                            },
                         {% endfor %}
                     ],
                     on: {
@@ -36,7 +36,7 @@ $$("content_right").addView({
                 {
                     view: 'label',
                     align: 'right',
-                    label: "{% trans 'Warning! if there are messages with an unknown status, the data is not updated' %}"
+                    label: "{{_("Warning! if there are messages with an unknown status, the data is not updated")|escapejs}}"
                 }
             ]
         },
@@ -63,44 +63,44 @@ $$("content_right").addView({
                                 columns: [
                                     {
                                         id: "period",
-                                        header: "{% trans 'Period' %}",
+                                        header: "{{_("Period")|escapejs}}",
                                         fillspace: true
                                     },
                                     {
                                         id: "messages_unknown",
-                                        header: "{% trans 'Unknowm status' %}",
+                                        header: "{{_("Unknowm status")|escapejs}}",
                                         fillspace: true
                                     },
                                     {
                                         id: "messages_fail",
-                                        header: "{% trans 'Not send' %}",
+                                        header: "{{_("Not send")|escapejs}}",
                                         fillspace: true
                                     },
                                     {
                                         id: "invoiced",
-                                        header: "{% trans 'Invoiced' %}",
+                                        header: "{{_("Invoiced")|escapejs}}",
                                         fillspace: true
                                     },
                                     {
                                         id: "to_be_invoiced",
-                                        header: "{% trans 'To be invoiced' %}",
+                                        header: "{{_("To be invoiced")|escapejs}}",
                                         fillspace: true
                                     },
                                     {
                                         id: "price_invoiced",
-                                        header: "{% trans 'Price invoiced' %}",
+                                        header: "{{_("Price invoiced")|escapejs}}",
                                         fillspace: true
                                     },
                                     {
                                         id: "price_to_be_invoiced",
-                                        header: "{% trans 'Price to be invoiced' %}",
+                                        header: "{{_("Price to be invoiced")|escapejs}}",
                                         fillspace: true
                                     },
                                     {
                                         id: "rating",
                                         header: "",
                                         template: function () {
-                                            return "<div class='webix_el_button'><button class='webixtype_base'>{% trans 'marks as billed' %}</button></div>";
+                                            return "<div class='webix_el_button'><button class='webixtype_base'>{{_("marks as billed")|escapejs}}</button></div>";
                                         },
                                         width: 200
                                     }
@@ -143,10 +143,10 @@ var marks_invoiced = function (datatable, id) {
     var sender = datatable.getItem(id.row).sender;
     var masks_billed = function (period) {
         webix.confirm({
-            title: "{% trans 'Confirmation' %}",
-            text: "{% trans 'Are you sure you want to mark these communications as billed?' %}",
-            ok: "{% trans 'Yes' %}",
-            cancel: "{% trans 'No' %}",
+            title: "{{_("Confirmation")|escapejs}}",
+            text: "{{_("Are you sure you want to mark these communications as billed?")|escapejs}}",
+            ok: "{{_("Yes")|escapejs}}",
+            cancel: "{{_("No")|escapejs}}",
             callback: function (result) {
                 var data = {
                     'period': period,
@@ -180,8 +180,8 @@ var marks_invoiced = function (datatable, id) {
 
     if (messages_unknown > 0) {
         webix.alert({
-            title: "{% trans 'Caution!' %}",
-            text: "{% trans 'There are unknown status for this communication' %}",
+            title: "{{_("Caution!")|escapejs}}",
+            text: "{{_("There are unknown status for this communication")|escapejs}}",
             callback: function () {
                 masks_billed(period);
             }