+++ CHANGELOG.rst
@@ -11,6 +11,34 @@ All notable changes to this project will be documented in this file.
 The format is based on KeepAChangelog_ and this project adheres to SemanticVersioning_.
 
 
+[1.0.0] - 2020-05-28
+++++++++++++++++++++
+
+Added
+~~~~~
+* Added maintainability badge
+* Added translations
+* Added email attachment link
+
+Changed
+~~~~~~~
+* `django-webix` min version v1.2.0
+* Better invoice area
+* Optimized count with exists
+
+Removed
+~~~~~~~
+* Removed old skebby gateway
+
+Fixed
+~~~~~
+* Fixed gateway utils without set sender as django app
+* Fixed filters
+* Fixed python3 compatibility
+* Fixed multiselect split
+* Fixed sender window typology autocomplete
+
+
 [0.3.6] - 2019-04-19
 ++++++++++++++++++++